requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
pydantic==2.10.4
PyJWT==2.10.1
bcrypt==4.2.1
python-dotenv==1.2.2
pytest==8.3.4
httpx==0.28.1
